requirements.txt
python-pptx
Pillow
pywin32 